refactor(domain): move pure rule modules under src/domain/rules/
Extract IO-free domain logic out of services/ into a co-located rules
package so services/ holds orchestration only:

- command-parser.ts moves verbatim
- stale-supersede.ts moves verbatim
- selectNextTasks lifts out of SchedulerService into a pure function in
  rules/scheduling.ts; the service class stays in services/ and just
  forwards (preserves the maxConcurrency option boundary)
- ExecutionPromptBuilder moves verbatim from services/execution/ into
  rules/execution-prompt.ts
- buildBranchName / buildCommitMessage / buildPullRequestTitle /
  withInstructionFooter lift out of ExecutionService private methods
  into rules/task-naming.ts

ExecutionService and EventDispatcher import the rules from their new
location. No behavior change; npm test 145 / 145.

Resolves #15
