+++ src/domain/rules/command-parser.ts
